Added /run endpoint for running kind code
Currently the endpoint compiles a hardcoded function and returns the
result as json. Next step will be to accept code as input and compile
it.

A few obstacles were encountered when trying to run the `kind` compiler
as a library. Most notably the fact that calls to `cerl` failed, which
turned out to be a dependency problem in `kind`: I hadn't added the
compiler package as a dependency and for that reason the modules in this
package like cerl were not included in the release.

Code runs locally now. To test, run `rebar3 run` and then in a separate
window:

```
curl localhost:2938/run
> {"response":"'Boolean/False'"}
```